+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added `ed2k` hash support for eDonkey/eMule/MLDonkey-compatible file hashes.
 - Added `tiger` Tiger Tree Hash support for Gnutella2/Direct Connect-compatible file hashes.
+- Added `btv2` hash support for BitTorrent v2 per-file `pieces root` hashes.
 
 ## 0.4.0 - 2026-04-30